(PUP-3548) Extract Runtime3Converter class from Runtime3Support module
The new Runtime3Converter class deals specifically with converting 4x
values into their 3x counterpart. Previously, this conversion was
built in to the Runtime3Support module.
